feat(xlsx-load): modifying web worker to load each sheet
