+++ src/workers/xlsxWorker.ts
@@ -3,12 +3,14 @@ import * as XLSX from "xlsx";
 
 self.onmessage = async (event) => {
   try {
-    const arrayBuffer = event.data; // We now receive ArrayBuffer directly
+    console.log(event.data)
+    const arrayBuffer = event.data.file; // We now receive ArrayBuffer directly
+    const targetSheet = event.data.sheetName;
     // Read workbook
     const workbook = XLSX.read(arrayBuffer);
     // Find "Stores" sheet
     const sheetName = workbook.SheetNames.find((name) =>
-      name.toLowerCase().includes("stores")
+      name.toLowerCase().includes(targetSheet)
     );
     if (!sheetName) {
       self.postMessage({ error: "Stores sheet not found!" });
@@ -18,7 +20,6 @@ self.onmessage = async (event) => {
     const sheet = workbook.Sheets[sheetName];
     const parsedData: Record<string, string>[] =
       XLSX.utils.sheet_to_json(sheet);
-    console.log(sheet);
     // Map data to Store objects
     const stores: Store[] = parsedData.map((store: Record<string, string>) => {
       return {
